requirements.txt
django
django-ipware
bandit
black
flake8
mypy
pylint
